requirements.txt
dspy-ai
dspy-ai[faiss-cpu]
langchain
openai
langchain_community
langchain_core
google-api-python-client
langchain-google-community
python-dotenv
googlesearch-python
apify 
apify-client
groq
fastapi
uvicorn